+++ packages/indexer/src/api/endpoints/execute/post-execute-transfer/v1.ts
@@ -13,7 +13,7 @@ const version = "v1";
 
 export const postExecuteTransferV1Options: RouteOptions = {
   description: "Transfer Tokens",
-  notes: "Use this endpoint to bulk transfer an array of NFTs."
+  notes: "Use this endpoint to bulk transfer an array of NFTs.",
   tags: ["api", "Trading"],
   plugins: {
     "hapi-swagger": {
