+++ packages/indexer/src/api/endpoints/search/get-search-collections/v2.ts
@@ -32,7 +32,9 @@ export const getSearchCollectionsV2Options: RouteOptions = {
     query: Joi.object({
       name: Joi.string()
         .lowercase()
-        .description("Lightweight search for collections that match a string. Can also search using contract address. Example: `bored` or `0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d`"),
+        .description(
+          "Lightweight search for collections that match a string. Can also search using contract address. Example: `bored` or `0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d`"
+        ),
       community: Joi.string()
         .lowercase()
         .description("Filter to a particular community. Example: `artblocks`"),
