+++ packages/indexer/src/api/endpoints/tokens/get-tokens/v7.ts
@@ -33,7 +33,8 @@ import { Sources } from "@/models/sources";
 import { Assets, ImageSize } from "@/utils/assets";
 import { CollectionSets } from "@/models/collection-sets";
 import { Collections } from "@/models/collections";
-import { getListedTokensFromES } from "@/api/endpoints/tokens/get-tokens/v6";
+import { getListedTokensFromES } from "@/api/endpoints/tokens";
+import { onchainMetadataProvider } from "@/metadata/providers/onchain-metadata-provider";
 
 const version = "v7";
 
@@ -318,7 +319,7 @@ export const getTokensV7Options: RouteOptions = {
                   value: JoiAttributeValue.description("Case sensitive."),
                   tokenCount: Joi.number(),
                   onSaleCount: Joi.number(),
-                  floorAskPrice: JoiPrice.allow(null).description("Can be null if no active asks."),
+                  floorAskPrice: Joi.number().unsafe().allow(null),
                   topBidValue: Joi.number().unsafe().allow(null),
                   createdAt: Joi.string(),
                 })
@@ -340,7 +341,8 @@ export const getTokensV7Options: RouteOptions = {
             floorAsk: {
               id: Joi.string().allow(null),
               price: JoiPrice.allow(null),
-              maker: Joi.string().lowercase().pattern(regex.address).allow(null),
+              // maker: Joi.string().lowercase().pattern(regex.address).allow(null),
+              maker: Joi.string().lowercase().allow(null),
               validFrom: Joi.number().unsafe().allow(null),
               validUntil: Joi.number().unsafe().allow(null),
               quantityFilled: Joi.number().unsafe().allow(null),
@@ -386,7 +388,6 @@ export const getTokensV7Options: RouteOptions = {
     let esTokens: any[] = [];
 
     const enableElasticsearchAsks =
-      config.enableElasticsearchAsks &&
       query.sortBy === "floorAskPrice" &&
       !["tokenName", "tokenSetId"].some((filter) => query[filter]);
 
@@ -427,9 +428,7 @@ export const getTokensV7Options: RouteOptions = {
                 'createdAt', ta.created_at,
                 'tokenCount', attributes.token_count,
                 'onSaleCount', attributes.on_sale_count,
-                'floorAskValue', attributes.floor_sell_value::TEXT,
-                'floorAskCurrency', attributes.floor_sell_currency,
-                'floorAskCurrencyValue', attributes.floor_sell_currency_value::TEXT,
+                'floorAskPrice', attributes.floor_sell_value::TEXT,
                 'topBidValue', attributes.top_buy_value::TEXT
               )
             )
@@ -728,8 +727,8 @@ export const getTokensV7Options: RouteOptions = {
           t.metadata,
           t.media,
           t.collection_id,
-          t.image_version,
           c.name AS collection_name,
+          t.image_version,
           con.kind,
           con.symbol,
           ${selectFloorData},
@@ -742,8 +741,6 @@ export const getTokensV7Options: RouteOptions = {
           t.supply,
           t.remaining_supply,
           extract(epoch from t.updated_at) AS t_updated_at,
-          t.metadata_disabled AS t_metadata_disabled,
-          c.metadata_disabled AS c_metadata_disabled,
           c.slug,
           c.creator,
           c.token_count,
@@ -1128,7 +1125,7 @@ export const getTokensV7Options: RouteOptions = {
         query.attributes ||
         query.tokenSetId ||
         query.rarity ||
-        (query.tokens && query.tokens.length > 1) ||
+        query.tokens ||
         (query.sortBy === "updatedAt" &&
           !(query.collectionsSetId || query.community || query.contract))
       ) {
@@ -1230,11 +1227,11 @@ export const getTokensV7Options: RouteOptions = {
       const rawResult = await redb.manyOrNone(baseQuery, query);
 
       /** Depending on how we sorted, we use that sorting key to determine the next page of results
-       Possible formats:
-       rarity_tokenid
-       floorAskPrice_tokenid
-       contract_tokenid
-       tokenid
+          Possible formats:
+            rarity_tokenid
+            floorAskPrice_tokenid
+            contract_tokenid
+            tokenid
        **/
       let continuation = null;
       if (rawResult.length === query.limit) {
@@ -1423,6 +1420,14 @@ export const getTokensV7Options: RouteOptions = {
           metadata.mediaOriginal = r.metadata.animation_original_url;
         }
 
+        if (!r.image && r.metadata?.image_original_url) {
+          r.image = onchainMetadataProvider.parseIPFSURI(r.metadata.image_original_url);
+        }
+
+        if (!r.media && r.metadata?.animation_original_url) {
+          r.media = onchainMetadataProvider.parseIPFSURI(r.metadata.animation_original_url);
+        }
+
         return {
           token: getJoiTokenObject(
             {
@@ -1501,35 +1506,20 @@ export const getTokensV7Options: RouteOptions = {
               owner: r.owner ? fromBuffer(r.owner) : null,
               attributes: query.includeAttributes
                 ? r.attributes
-                  ? await Promise.all(
-                      _.map(r.attributes, async (attribute) => ({
-                        key: attribute.key,
-                        kind: attribute.kind,
-                        value: attribute.value,
-                        tokenCount: attribute.tokenCount,
-                        onSaleCount: attribute.onSaleCount,
-                        floorAskPrice: attribute.floorAskValue
-                          ? await getJoiPriceObject(
-                              {
-                                gross: {
-                                  amount: String(
-                                    attribute.floorAskCurrencyValue ?? attribute.floorAskValue
-                                  ),
-                                  nativeAmount: String(attribute.floorAskValue),
-                                },
-                              },
-                              attribute.floorAskCurrency
-                                ? _.replace(attribute.floorAskCurrency, "\\x", "0x")
-                                : Sdk.Common.Addresses.Native[config.chainId],
-                              query.displayCurrency
-                            )
-                          : null,
-                        topBidValue: attribute.topBidValue
-                          ? formatEth(attribute.topBidValue)
-                          : attribute.topBidValue,
-                        createdAt: new Date(attribute.createdAt).toISOString(),
-                      }))
-                    )
+                  ? _.map(r.attributes, (attribute) => ({
+                      key: attribute.key,
+                      kind: attribute.kind,
+                      value: attribute.value,
+                      tokenCount: attribute.tokenCount,
+                      onSaleCount: attribute.onSaleCount,
+                      floorAskPrice: attribute.floorAskPrice
+                        ? formatEth(attribute.floorAskPrice)
+                        : attribute.floorAskPrice,
+                      topBidValue: attribute.topBidValue
+                        ? formatEth(attribute.topBidValue)
+                        : attribute.topBidValue,
+                      createdAt: new Date(attribute.createdAt).toISOString(),
+                    }))
                   : []
                 : undefined,
               mintStages: r.mint_stages
@@ -1551,6 +1541,147 @@ export const getTokensV7Options: RouteOptions = {
             r.t_metadata_disabled,
             r.c_metadata_disabled
           ),
+          market: {
+            floorAsk: {
+              id: r.floor_sell_id,
+              price: r.floor_sell_id
+                ? await getJoiPriceObject(
+                    {
+                      gross: {
+                        amount: r.floor_sell_currency_value ?? r.floor_sell_value,
+                        nativeAmount: r.floor_sell_value,
+                      },
+                    },
+                    floorAskCurrency,
+                    query.displayCurrency
+                  )
+                : null,
+              maker: r.floor_sell_maker ? fromBuffer(r.floor_sell_maker) : null,
+              validFrom: r.floor_sell_value ? r.floor_sell_valid_from : null,
+              validUntil: r.floor_sell_value ? r.floor_sell_valid_to : null,
+              quantityFilled:
+                query.includeQuantity && r.floor_sell_value
+                  ? r.floor_sell_quantity_filled
+                  : undefined,
+              quantityRemaining:
+                query.includeQuantity && r.floor_sell_value
+                  ? r.floor_sell_quantity_remaining
+                  : undefined,
+              dynamicPricing,
+              source: getJoiSourceObject(floorSellSource),
+            },
+            topBid: query.includeTopBid
+              ? {
+                  id: r.top_buy_id,
+                  price: r.top_buy_value
+                    ? await getJoiPriceObject(
+                        {
+                          net: {
+                            amount: query.normalizeRoyalties
+                              ? r.top_buy_currency_normalized_value ?? r.top_buy_value
+                              : r.top_buy_currency_value ?? r.top_buy_value,
+                            nativeAmount: query.normalizeRoyalties
+                              ? r.top_buy_normalized_value ?? r.top_buy_value
+                              : r.top_buy_value,
+                          },
+                          gross: {
+                            amount: r.top_buy_currency_price ?? r.top_buy_price,
+                            nativeAmount: r.top_buy_price,
+                          },
+                        },
+                        topBidCurrency,
+                        query.displayCurrency
+                      )
+                    : null,
+                  maker: r.top_buy_maker ? fromBuffer(r.top_buy_maker) : null,
+                  validFrom: r.top_buy_valid_from,
+                  validUntil: r.top_buy_value ? r.top_buy_valid_until : null,
+                  source: getJoiSourceObject(topBuySource),
+                  feeBreakdown: feeBreakdown,
+                }
+              : undefined,
+          },
+          updatedAt: new Date(r.t_updated_at * 1000).toISOString(),
+        };
+
+        return {
+          token: {
+            chainId: config.chainId,
+            contract,
+            tokenId,
+            name: r.name,
+            description: r.description,
+            image: Assets.getResizedImageUrl(r.image),
+            imageSmall: Assets.getResizedImageUrl(r.image, ImageSize.small),
+            imageLarge: Assets.getResizedImageUrl(r.image, ImageSize.large),
+            metadata: Object.values(metadata).every((el) => el === undefined)
+              ? undefined
+              : metadata,
+            media: r.media,
+            kind: r.kind,
+            isFlagged: Boolean(Number(r.is_flagged)),
+            isSpam: Number(r.t_is_spam) > 0 || Number(r.c_is_spam) > 0,
+            metadataDisabled:
+              Boolean(Number(r.t_metadata_disabled)) || Boolean(Number(r.c_metadata_disabled)),
+            lastFlagUpdate: r.last_flag_update ? new Date(r.last_flag_update).toISOString() : null,
+            lastFlagChange: r.last_flag_change ? new Date(r.last_flag_change).toISOString() : null,
+            supply: !_.isNull(r.supply) ? r.supply : null,
+            remainingSupply: !_.isNull(r.remaining_supply) ? r.remaining_supply : null,
+            rarity: r.rarity_score,
+            rarityRank: r.rarity_rank,
+            collection: {
+              id: r.collection_id,
+              name: r.collection_name,
+              image: Assets.getLocalAssetsLink(r.collection_image),
+              slug: r.slug,
+              symbol: r.symbol,
+              creator: r.creator ? fromBuffer(r.creator) : null,
+              tokenCount: r.token_count,
+              metadataDisabled: Boolean(Number(r.c_metadata_disabled)),
+            },
+            lastSale:
+              query.includeLastSale && r.last_sale_currency
+                ? await getJoiSaleObject({
+                    prices: {
+                      gross: {
+                        amount: r.last_sale_currency_price ?? r.last_sale_price,
+                        nativeAmount: r.last_sale_price,
+                        usdAmount: r.last_sale_usd_price,
+                      },
+                    },
+                    fees: {
+                      royaltyFeeBps: r.last_sale_royalty_fee_bps,
+                      marketplaceFeeBps: r.last_sale_marketplace_fee_bps,
+                      paidFullRoyalty: r.last_sale_paid_full_royalty,
+                      royaltyFeeBreakdown: r.last_sale_royalty_fee_breakdown,
+                      marketplaceFeeBreakdown: r.last_sale_marketplace_fee_breakdown,
+                    },
+                    currencyAddress: r.last_sale_currency,
+                    timestamp: r.last_sale_timestamp,
+                    orderSourceId: r.last_sale_order_source_id_int,
+                    fillSourceId: r.last_sale_fill_source_id,
+                  })
+                : undefined,
+            owner: r.owner ? fromBuffer(r.owner) : null,
+            attributes: query.includeAttributes
+              ? r.attributes
+                ? _.map(r.attributes, (attribute) => ({
+                    key: attribute.key,
+                    kind: attribute.kind,
+                    value: attribute.value,
+                    tokenCount: attribute.tokenCount,
+                    onSaleCount: attribute.onSaleCount,
+                    floorAskPrice: attribute.floorAskPrice
+                      ? formatEth(attribute.floorAskPrice)
+                      : attribute.floorAskPrice,
+                    topBidValue: attribute.topBidValue
+                      ? formatEth(attribute.topBidValue)
+                      : attribute.topBidValue,
+                    createdAt: new Date(attribute.createdAt).toISOString(),
+                  }))
+                : []
+              : undefined,
+          },
           market: {
             floorAsk: {
               id: r.floor_sell_id,
