+++ packages/indexer/src/api/endpoints/tokens/index.ts
@@ -5,7 +5,6 @@ export * from "@/api/endpoints/tokens/get-tokens/v4";
 export * from "@/api/endpoints/tokens/get-tokens/v5";
 export * from "@/api/endpoints/tokens/get-tokens/v6";
 export * from "@/api/endpoints/tokens/get-tokens/v7";
-export * from "@/api/endpoints/tokens/get-tokens/v8";
 export * from "@/api/endpoints/tokens/get-tokens-bootstrap/v1";
 export * from "@/api/endpoints/tokens/get-tokens-details/v2";
 export * from "@/api/endpoints/tokens/get-tokens-details/v3";
